new sqlalchemy model for user-annotation-state
